+++ flutter_reactive_form/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Initialize the reactive form.
+
+## 0.0.4
+
+* Modify the return type in ValidationFunc.